backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0
alembic==1.14.0
pydantic-settings==2.6.1
pydantic==2.9.2
python-multipart==0.0.19
aiogram==3.15.0
google-generativeai==0.8.3
aiofiles==24.1.0
python-dateutil==2.9.0.post0
httpx==0.28.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
